backend/requirements.txt
#
# This file is autogenerated by pip-compile with python 3.8
# To update, run:
#
#    pip-compile requirements.in
#
appdirs==1.4.4
    # via black
asgiref==3.4.1
    # via django
black==21.7b0
    # via -r requirements.in
certifi==2021.5.30
    # via requests
charset-normalizer==2.0.4
    # via requests
click==8.0.1
    # via black
colorama==0.4.4
    # via click
coreapi==2.3.3
    # via djangorestframework-stubs
coreschema==0.0.4
    # via coreapi
django==3.2.5
    # via
    #   -r requirements.in
    #   django-cors-headers
    #   django-extensions
    #   django-filter
    #   django-stubs
    #   django-stubs-ext
    #   djangorestframework
    #   djangorestframework-simplejwt
django-cors-headers==3.7.0
    # via -r requirements.in
django-extensions==3.1.3
    # via -r requirements.in
django-filter==2.4.0
    # via -r requirements.in
django-stubs==1.8.0
    # via
    #   -r requirements.in
    #   djangorestframework-stubs
django-stubs-ext==0.2.0
    # via
    #   -r requirements.in
    #   django-stubs
djangorestframework==3.12.4
    # via
    #   -r requirements.in
    #   djangorestframework-simplejwt
djangorestframework-simplejwt==4.7.2
    # via -r requirements.in
djangorestframework-stubs==1.4.0
    # via -r requirements.in
flake8==3.9.2
    # via -r requirements.in
idna==3.2
    # via requests
itypes==1.2.0
    # via coreapi
jinja2==3.0.1
    # via coreschema
lxml==4.6.3
    # via -r requirements.in
markdown==3.3.4
    # via -r requirements.in
markupsafe==2.0.1
    # via jinja2
mccabe==0.6.1
    # via flake8
mypy==0.910
    # via
    #   -r requirements.in
    #   django-stubs
    #   djangorestframework-stubs
mypy-extensions==0.4.3
    # via
    #   black
    #   mypy
pathspec==0.9.0
    # via black
psycopg2==2.9.1
    # via -r requirements.in
pycodestyle==2.7.0
    # via flake8
pyflakes==2.3.1
    # via flake8
pyjwt==2.1.0
    # via djangorestframework-simplejwt
pytz==2021.1
    # via django
regex==2021.7.6
    # via black
requests==2.26.0
    # via
    #   coreapi
    #   djangorestframework-stubs
sqlparse==0.4.1
    # via django
toml==0.10.2
    # via mypy
tomli==1.0.4
    # via black
typing-extensions==3.10.0.0
    # via
    #   django-stubs
    #   djangorestframework-stubs
    #   mypy
uritemplate==3.0.1
    # via coreapi
urllib3==1.26.6
    # via requests
